+++ product_code/operation_service/impl/src/test/scala/de/upb/cs/uc4/operation/impl/OperationStateSpec.scala
@@ -3,7 +3,7 @@ package de.upb.cs.uc4.operation.impl
 import akka.actor.testkit.typed.scaladsl.ScalaTestWithActorTestKit
 import akka.persistence.typed.PersistenceId
 import de.upb.cs.uc4.operation.impl.actor.{ OperationDatabaseBehaviour, WatchlistWrapper }
-import de.upb.cs.uc4.operation.impl.commands.{ AddToWatchlist, GetWatchlist, RemoveFromWatchlist }
+import de.upb.cs.uc4.operation.impl.commands.{ AddToWatchlist, ClearWatchlist, GetWatchlist, RemoveFromWatchlist }
 import de.upb.cs.uc4.shared.server.messages.{ Accepted, Confirmation }
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
@@ -95,6 +95,26 @@ class OperationStateSpec extends ScalaTestWithActorTestKit(s"""
 
       val probe3 = createTestProbe[Confirmation]()
       ref ! RemoveFromWatchlist(opId, probe3.ref)
+      probe3.expectMessageType[Accepted]
+
+      val probe4 = createTestProbe[WatchlistWrapper]()
+      ref ! GetWatchlist(probe4.ref)
+      probe4.expectMessage(WatchlistWrapper(Seq()))
+    }
+
+    "clear the watchlist" in {
+      val opId = "operation1"
+      val ref = spawn(OperationDatabaseBehaviour.create(PersistenceId("fake-type-hint", "fake-id-2")))
+      val probe1 = createTestProbe[Confirmation]()
+      ref ! AddToWatchlist(opId, probe1.ref)
+      probe1.expectMessageType[Accepted]
+
+      val probe2 = createTestProbe[WatchlistWrapper]()
+      ref ! GetWatchlist(probe2.ref)
+      probe2.expectMessage(WatchlistWrapper(Seq(opId)))
+
+      val probe3 = createTestProbe[Confirmation]()
+      ref ! ClearWatchlist("username", probe3.ref)
       probe3.expectMessageType[Accepted]
 
       val probe4 = createTestProbe[WatchlistWrapper]()